rearanged location service
